fix(sim): simulation deep-review remediation (epic Luciferase-0frcy)

+++ simulation/src/main/java/com/hellblazer/luciferase/simulation/persistence/RecoveredState.java
@@ -30,7 +30,7 @@ import java.util.Objects;
 public record RecoveredState(
     CheckpointMetadata checkpoint,
     List<Map<String, Object>> events,
-    int totalEventsReplayed,
+    long totalEventsReplayed,
     int skippedEvents
 ) {
